Melhorei a organizaçao das funçoes e pastas para melhor controle de versoes, codigo mais limpo e muito mais... :)
